+++ next.config.js
@@ -1,7 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
     images: {
-        domains: ["demo.wpthemego.com"]
+        domains: ["demo.wpthemego.com", "s.s-bol.com"]
     }
 }