+++ src/app/api/books/route.ts
@@ -1,5 +1,66 @@
-import {BOOKS_DATA} from "@/utils/db.util";
 
+const BOOKS_DATA: ReadonlyArray<IBook> = [
+    {
+        id: 1,
+        name: "An Italian Holiday",
+        author: "mamadfar",
+        price: 105,
+        summary: "Style Code Live is a daily, live show where style enthusiasts can connect, chat, shop, and get the inside scoop on the latest fashion and beauty trends.",
+        stock: 20,
+        rate: 3.5,
+        categories: ["Art", "Literature", "Romance"],
+        description: "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur. Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo",
+        imageUrl: "https://demo.wpthemego.com/themes/sw_emarket/layout20/wp-content/uploads/2020/07/b4.png"
+    },
+    {
+        id: 2,
+        name: "Magni Dolores Eos",
+        author: "mamadfar",
+        price: 110,
+        summary: "Style Code Live is a daily, live show where style enthusiasts can connect, chat, shop, and get the inside scoop on the latest fashion and beauty trends.",
+        stock: 5,
+        rate: 2.3,
+        categories: ["Art", "Fantasy", "History", "Literature", "Romance"],
+        description: "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur. Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo",
+        imageUrl: "https://demo.wpthemego.com/themes/sw_emarket/layout20/wp-content/uploads/2020/07/b14.png"
+    },
+    {
+        id: 3,
+        name: "Natur Aut Odit Aut",
+        author: "mamadfar",
+        price: 130,
+        summary: "Style Code Live is a daily, live show where style enthusiasts can connect, chat, shop, and get the inside scoop on the latest fashion and beauty trends.",
+        stock: 7,
+        rate: 4.1,
+        categories: ["Fantasy", "History", "Literature", "Romance"],
+        description: "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur. Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo",
+        imageUrl: "https://demo.wpthemego.com/themes/sw_emarket/layout20/wp-content/uploads/2020/07/b13.png"
+    },
+    {
+        id: 4,
+        name: "Julia’s Story",
+        author: "mamadfar",
+        price: 75,
+        summary: "Style Code Live is a daily, live show where style enthusiasts can connect, chat, shop, and get the inside scoop on the latest fashion and beauty trends.",
+        stock: 30,
+        rate: 1.3,
+        categories: ["Art", "History", "Literature", "Romance"],
+        description: "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur. Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo",
+        imageUrl: "https://demo.wpthemego.com/themes/sw_emarket/layout20/wp-content/uploads/2020/07/b5.png"
+    },
+    {
+        id: 5,
+        name: "Ut Enim Ad Mini",
+        author: "karim",
+        price: 180,
+        summary: "Style Code Live is a daily, live show where style enthusiasts can connect, chat, shop, and get the inside scoop on the latest fashion and beauty trends.",
+        stock: 2,
+        rate: 3.6,
+        categories: ["Art", "History", "Romance"],
+        description: "Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum exercitationem ullam corporis suscipit laboriosam, nisi ut aliquid ex ea commodi consequatur. Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo",
+        imageUrl: "https://demo.wpthemego.com/themes/sw_emarket/layout20/wp-content/uploads/2020/07/b7.png"
+    }
+]
 export async function GET(req: Request, res: Response) {
     const body = JSON.stringify({results: [...BOOKS_DATA]});
     return new Response(body, {