+++ src/services/books.service.ts
@@ -1,5 +1,9 @@
 import API from "@/config/API";
 
 export const getBooksService = () => {
-        return API.get<IResponse<ReadonlyArray<IBook>>>("http://localhost:3000/api/books");
+        return API.get<ReadonlyArray<IBook>>("/data/books");
+}
+
+export const getBookService = (bookId: number | string, signal?: AbortSignal) => {
+        return API.get<IBookResponse<IBook>>(`/books/${bookId}`, {signal});
 }