+++ src/types/books.d.ts
@@ -1,12 +1,24 @@
+// interface IBook {
+//     id: number | string;
+//     name: string;
+//     author: string;
+//     summary: string;
+//     description: string;
+//     imageUrl: string;
+//     stock: number;
+//     rate: number;
+//     price: number;
+//     categories: string[];
+// }
 interface IBook {
-    id: number | string;
-    name: string;
+    ISBN: number;
+    title: string;
     author: string;
     summary: string;
-    description: string;
-    imageUrl: string;
-    stock: number;
-    rate: number;
-    price: number;
-    categories: string[];
+    image: string;
+    price: {
+        currency: string,
+        value: number,
+        displayValue: string
+    };
 }